:sparkles: Bot Navigation
Bots will now use navigation regions and agents to traverse the map to targets. The current target is the player at all times, this will change to a dynamic system later on. Bots will try to avoid each other and will come within 150 pixels of the player to hover around them.